Move TibberDataCoordinator to separate module
